+++ packages/widget/src/pages/SwapExecutionPage/SwapExecutionPageRouteDetailed.tsx
@@ -28,6 +28,7 @@ const operationTypeToIcon: operationTypeToIcon = {
   [OperationType.hyperlaneTransfer]: <SwapExecutionBridgeIcon width={34} />,
   [OperationType.opInitTransfer]: <SwapExecutionBridgeIcon width={34} />,
   [OperationType.goFastTransfer]: <SwapExecutionBridgeIcon width={34} />,
+  [OperationType.stargateTransfer]: <SwapExecutionBridgeIcon width={34} />,
   // send icon
   [OperationType.bankSend]: <SwapExecutionSendIcon width={34} />,
 };
@@ -42,6 +43,7 @@ const operationTypeToSimpleOperationType = {
   opInitTransfer: "Bridged",
   bankSend: "Sent",
   goFastTransfer: "Bridged",
+  stargateTransfer: "Bridged",
 };
 
 type tooltipMap = Record<number, boolean>;